pyproject.toml
[project]
name = "zuup-forge"
version = "0.1.0"
description = "The Platform That Builds Platforms"
readme = "readme.md"
license = "Apache-2.0"
requires-python = ">=3.11"
authors = [{name = "Aldrich K. Wooden, Sr."}]
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "httpx>=0.27",
    "werkzeug>=1.0.1",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23", "ruff>=0.4", "mypy>=1.10"]
prod = ["gunicorn>=22.0", "opentelemetry-api>=1.24", "opentelemetry-sdk>=1.24"]
ml = ["sentence-transformers>=3.0", "trl>=0.8"]

[project.scripts]
forge = "forge.cli.main:main"

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
exclude = ["api*", "public*"]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM"]
ignore = ["N999"]  # project name zuup-forge is intentional

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.11"
strict = true
